Rename sublime_lib.constants to flags
Also remove the unused key names.
