--- plugins_/lib/sublime_lib/constants.py
@@ -1,91 +1,0 @@
-import logging
-import sublime
-
-l = logging.getLogger(__name__)
-
-KEY_UP                  = "up"
-KEY_DOWN                = "down"
-KEY_RIGHT               = "right"
-KEY_LEFT                = "left"
-KEY_INSERT              = "insert"
-KEY_HOME                = "home"
-KEY_END                 = "end"
-KEY_PAGEUP              = "pageup"
-KEY_PAGEDOWN            = "pagedown"
-KEY_BACKSPACE           = "backspace"
-KEY_DELETE              = "delete"
-KEY_TAB                 = "tab"
-KEY_ENTER               = "enter"
-KEY_PAUSE               = "pause"
-KEY_ESCAPE              = "escape"
-KEY_SPACE               = "space"
-KEY_KEYPAD0             = "keypad0"
-KEY_KEYPAD1             = "keypad1"
-KEY_KEYPAD2             = "keypad2"
-KEY_KEYPAD3             = "keypad3"
-KEY_KEYPAD4             = "keypad4"
-KEY_KEYPAD5             = "keypad5"
-KEY_KEYPAD6             = "keypad6"
-KEY_KEYPAD7             = "keypad7"
-KEY_KEYPAD8             = "keypad8"
-KEY_KEYPAD9             = "keypad9"
-KEY_KEYPAD_PERIOD       = "keypad_period"
-KEY_KEYPAD_DIVIDE       = "keypad_divide"
-KEY_KEYPAD_MULTIPLY     = "keypad_multiply"
-KEY_KEYPAD_MINUS        = "keypad_minus"
-KEY_KEYPAD_PLUS         = "keypad_plus"
-KEY_KEYPAD_ENTER        = "keypad_enter"
-KEY_CLEAR               = "clear"
-KEY_F1                  = "f1"
-KEY_F2                  = "f2"
-KEY_F3                  = "f3"
-KEY_F4                  = "f4"
-KEY_F5                  = "f5"
-KEY_F6                  = "f6"
-KEY_F7                  = "f7"
-KEY_F8                  = "f8"
-KEY_F9                  = "f9"
-KEY_F10                 = "f10"
-KEY_F11                 = "f11"
-KEY_F12                 = "f12"
-KEY_F13                 = "f13"
-KEY_F14                 = "f14"
-KEY_F15                 = "f15"
-KEY_F16                 = "f16"
-KEY_F17                 = "f17"
-KEY_F18                 = "f18"
-KEY_F19                 = "f19"
-KEY_F20                 = "f20"
-KEY_SYSREQ              = "sysreq"
-KEY_BREAK               = "break"
-KEY_CONTEXT_MENU        = "context_menu"
-KEY_BROWSER_BACK        = "browser_back"
-KEY_BROWSER_FORWARD     = "browser_forward"
-KEY_BROWSER_REFRESH     = "browser_refresh"
-KEY_BROWSER_STOP        = "browser_stop"
-KEY_BROWSER_SEARCH      = "browser_search"
-KEY_BROWSER_FAVORITES   = "browser_favorites"
-KEY_BROWSER_HOME        = "browser_home"
-
-# The following available add_region styles are taken from the API documentation:
-# http://www.sublimetext.com/docs/3/api_reference.html#sublime.View
-# unfortunately, the `sublime` module doesn't encapsulate them for easy reference
-# so we hardcode them here.
-# TODO use `enum` lib
-ADD_REGION_STYLE_NAMES = {
-    'DRAW_EMPTY', 'HIDE_ON_MINIMAP', 'DRAW_EMPTY_AS_OVERWRITE', 'DRAW_NO_FILL',
-    'DRAW_NO_OUTLINE', 'DRAW_SOLID_UNDERLINE', 'DRAW_STIPPLED_UNDERLINE',
-    'DRAW_SQUIGGLY_UNDERLINE', 'HIDDEN', 'PERSISTENT',
-}
-
-
-def style_flags_from_list(style_list):
-    style_flags = 0
-    if not isinstance(style_list, list):
-        raise TypeError("Expected style *list*")
-    for style in style_list:
-        if style in ADD_REGION_STYLE_NAMES:
-            style_flags |= getattr(sublime, style)
-        else:
-            l.debug("Style flag %r doesn't exist", style)
-    return style_flags
