+++ plugins_/settings/__init__.py
@@ -6,7 +6,7 @@ import sublime
 import sublime_plugin
 
 from ..lib import get_setting
-from ..lib.sublime_lib.constants import style_flags_from_list
+from ..lib.flags import style_flags_from_list
 from ..lib.weakmethod import WeakMethodProxy
 
 from .region_math import (VALUE_SCOPE, KEY_SCOPE, KEY_COMPLETIONS_SCOPE,
